[Agent] split GameEngine tests
